+++ RELEASES.md
@@ -1,10 +1,18 @@
+# Astra Monitor 3 - TO BE RELEASED
+
+### New features
+- Swap Memory info expanded with a popup menu to show more details, including swap devices and usage per device
+
+### Bug fixes
+- __Swap__-ped from 'free' to '/proc/meminfo' to get swap info. [[#6](https://github.com/AstraExt/astra-monitor/issues/6)]
+
 # Astra Monitor 2 - January 12 2024
 
 ### New features
 - Added an option to change Theme Style (Light/Dark) to better fit your desktop environment's theme
 
 ### Bug fixes
-- Sensor header width not consistant [#1](https://github.com/AstraExt/astra-monitor/issues/1)
+- Sensor header width not consistant [[#1](https://github.com/AstraExt/astra-monitor/issues/1)]
 - Sensor menu not centered
 - Sensor menu missing footer utility buttons [[#4](https://github.com/AstraExt/astra-monitor/issues/4)]
 - Icon madness fixed. Now icons are more unified and support light and dark themes